tellBaby/requirements.txt
Django==2.0
PyMySQL
Pillow
django-import-export==0.6.1